FEATURE: Add scraper for Carfolio makes and models
This adds a scraper for Carfolio's makes and models.
This iterates thru Carfolio's specifications page and grabs the name,
country, and link for each manufacturer (make), then uses that to grab
the model names, years (if available), and link for each make.

This does not yet grab any details about each model.

There is some refactoring to be done on this code.