FEATURE: Use global Reqwest client, add timings
This changes our HTTP access to use a single client, rather than
instantiating a new one per request.

This also adds some simply timing output for general benchmarking.

Finally, converts Model `year` to be a string, so that it can be an
empty string, as opposed to an `Option<u16>`.