REFACTOR: Further cleanup HTML scraping bits
This creates a Page struct that will request it's HTML upon
instantiation, and hold the response in it's fields.

Then, further abstract the HTML bits so that the `extract_*` functions
can generally just be a very clear single function call to get their
values.

Finally, re-organize the `scrape` function a little bit to get
everything to work again, although this is going to change again soon
when optimized for asynchronous processing.