REFACTOR: Fix error messages and logging
This makes error messages not log on creation anymore, since that
doesn't make sense if the error is handled.

This makes it so that the logging does not happen automatically.
Instead, logging should happen explicitly and intentionally. It simply
calls the `std::fmt::Display::fmt` function, which prints the Error
message out nicely.