FEATURE: Scrape car details
The adds code to scrape the detailed car pages.
Each of these pages contain a table with the specifications for the car.
This new code will grab this table and put all the specifications into a
BTreeMap.

The keys of the BTreeMap are the name of the specification, e.g.
"horsepower". These are lowercased, and underscored for uniformity.

The values of the BTreeMap are the specifications themselves, e.g.
"400", for horsepower. Note that all values are strings, but trimmed and
sanitized.

Also note that some rows may include multiple `td`s in them. In these
cases, the code is currently just grabbing the first one, since that
appears to be the metric value. The second is the imperial value. We can
live with just the metric value. If it is noticed that this is not
always the case, then we can deal with that later.