FEATURE: Start scraping car details
This creates a module for vehicle under carfolio that contains the
detailed specifications of a vehicle.

The HTML table for car specifications is scraped and each field is
individually parsed.

Each field is an `Option<T>` because there isn't any guarantee that any
field will be present.

So far, this is just progress on maybe 1/3 of the available fields. With
the rest still to be coded, or possibly, ignored.