FEATURE: Change logging levels for messages, add skiplists for fields
This changes the logging level of some really detailed things so that
INFO isn't so packed all the time.

This also adds some skiplist in order to allow skipping of certain `th`
fields from the specifications table, and also fields to ignore in the
warning about unused specifications when populating the Vehicle struct.